fix order_points()
Fixed the `order_points()` function for the semi-pathological case where you have a rectangle that is skewed so that the bottom 2 points are to the left of the top 2 points.  For instance, if you have a rectangle with coordinates `[[5, 1], [6,1], [2, 3], [1,3]]` (which is in clockwise-order) the old function will give you: `[[1, 3], [5, 1], [6, 1], [2, 3]]` which has the bottom-left point leading.