Fixed helper crashing on face_utils.visualize_facial_landmarks(image, shape)
Hello ! this is the first time i do this, so bear with me if i'm a little clueless.

So basically, i tried to follow this tutorial https://www.pyimagesearch.com/2017/04/10/detect-eyes-nose-lips-jaw-dlib-opencv-python/ and it crashed on the part showing the transparent face features.

I debugged it, and turns out it was this function. colors (line 83) was manipulated like it started from 1, not from 0; constantly throwing an out of range error.