fix last frame being None
When FileVideoStream reaches the end of the file, it takes one more frame (which is NoneType) and tries to transform and add it to the queue. This has also been noted in issue #251 . This causes an error if the transform expects a valid frame, or if the main script does not expect the last frame in the Queue to be NoneType.