Changing 'angle' in rotate_bound() from negative to positive to unify it with rotate() funciton.
Since definition of rotation matrix around specific axis is that the positive angle is counter clockwise rotation, I guess it should be like this in rotate_bound() function. Additionally consider that in rotate() function angle is counter clockwise, so making it the same in rotate_bound() would unify the library.