Strong Typed var that was causing error.
When used on Python3, this utility broke with the error "TypeError: Can't parse 'center'. Sequence item with index 0 has a wrong type". I debugged it and found that the error was the absence of typing on the variable eyesCenter, so I strong typed the var that was causing the error.